+++ QGoodWindow/Examples/GoodShowCase/main.cpp
@@ -27,9 +27,6 @@ SOFTWARE.
 
 int main(int argc, char *argv[])
 {
-#ifdef Q_OS_ANDROID
-    QApplication::setAttribute(Qt::AA_EnableHighDpiScaling);
-#endif
     QApplication a(argc, argv);
 
     MainWindow mw;