+++ QGoodWindow/Examples/GoodShowCase/mainwindow.cpp
@@ -26,26 +26,24 @@ SOFTWARE.
 
 MainWindow::MainWindow(QWidget *parent) : QGoodWindow(parent)
 {
-#ifdef QGOODWINDOW
-    title_bar = new TitleBar(pixelRatio(), this);
-
-    frame = new QFrame(this);
+    m_dark = false;
 
     m_frame_style = QString("QFrame {background-color: %0; border: %1;}");
 
     m_color_str = "#FFFFFF";
 
-#ifndef Q_OS_MAC
-    frame->setStyleSheet(m_frame_style.arg(m_color_str).arg("1px solid #1883D7"));
-#else
-    frame->setStyleSheet(m_frame_style.arg(m_color_str).arg("none"));
-#endif
+    QString border_str = "none";
+
+#ifdef QGOODWINDOW
+    title_bar = new TitleBar(pixelRatio(), this);
+
+    frame = new QFrame(this);
 
     QVBoxLayout *layout = new QVBoxLayout(frame);
     layout->setMargin(0);
     layout->setSpacing(0);
     layout->addWidget(title_bar);
-    layout->addStretch(1);
+    layout->addStretch();
 
     setCentralWidget(frame);
 
@@ -60,42 +58,52 @@ MainWindow::MainWindow(QWidget *parent) : QGoodWindow(parent)
 
     connect(this, &MainWindow::windowIconChanged, this, [=](const QIcon &icon){
         if (!icon.isNull())
-            title_bar->setIcon(icon.pixmap(16, 16));
+        {
+            const int pix_size = qCeil(16 * pixelRatio());
+            title_bar->setIcon(icon.pixmap(pix_size, pix_size));
+        }
     });
-#else
-    QString m_frame_style = QString("QFrame {background-color: #FFFFFF;}");
 
-    QFrame *frame = new QFrame(this);
-    frame->setStyleSheet(m_frame_style);
+#ifndef Q_OS_MAC
+    if (windowState().testFlag(Qt::WindowNoState))
+        border_str = "1px solid #1883D7";
+#endif
+
+#else
+    frame = new QFrame(this);
 
     setCentralWidget(frame);
 #endif
-    resize(qRound(640 * pixelRatio()), qRound(480 * pixelRatio()));
 
-    move(QGuiApplication::primaryScreen()->availableGeometry().center() - rect().center());
+    frame->setStyleSheet(m_frame_style.arg(m_color_str).arg(border_str));
+
+    QShortcut *shortcut1 = new QShortcut(QKeySequence(Qt::Key_S), this);
+
+    connect(shortcut1, &QShortcut::activated, this, [=]{
+        m_dark = !m_dark;
 
 #ifdef QGOODWINDOW
-    QShortcut *shortcut = new QShortcut(QKeySequence(Qt::Key_S), this);
+        title_bar->setDarkMode(m_dark);
+#endif
 
-    connect(shortcut, &QShortcut::activated, this, [=]{
-        title_bar->setDarkMode(!title_bar->dark());
-
-        if (!title_bar->dark())
+        if (!m_dark)
             m_color_str = "#FFFFFF";
         else
             m_color_str = "#000000";
 
-#ifndef Q_OS_MAC
-        frame->setStyleSheet(m_frame_style.arg(m_color_str)
-                             .arg((isFullScreen() || isMaximized()) ? "none" : "1px solid #1883D7"));
-#else
-        frame->setStyleSheet(m_frame_style.arg(m_color_str).arg("none"));
-#endif
-    });
-#endif
-    QShortcut *shortcut1 = new QShortcut(QKeySequence(Qt::Key_F), this);
+        QString border_str = "none";
 
-    connect(shortcut1, &QShortcut::activated, this, [=]{
+#if defined QGOODWINDOW && !defined Q_OS_MAC
+        if (windowState().testFlag(Qt::WindowNoState))
+            border_str = "1px solid #1883D7";
+#endif
+
+        frame->setStyleSheet(m_frame_style.arg(m_color_str).arg(border_str));
+    });
+
+    QShortcut *shortcut2 = new QShortcut(QKeySequence(Qt::Key_F), this);
+
+    connect(shortcut2, &QShortcut::activated, this, [=]{
         if (!isFullScreen())
             showFullScreen();
         else
@@ -107,10 +115,25 @@ MainWindow::MainWindow(QWidget *parent) : QGoodWindow(parent)
 
     setWindowIcon(p);
 
-#ifdef QGOODWINDOW
     setWindowTitle("Good Window - Press S to toggle theme - Press F to toggle fullscreen!");
-#else
-    setWindowTitle("Good Window - Press F to toggle fullscreen!");
+
+    resize(qRound(640 * pixelRatio()), qRound(480 * pixelRatio()));
+
+    move(QGuiApplication::primaryScreen()->availableGeometry().center() - rect().center());
+
+#ifdef Q_OS_WIN
+    QTimer::singleShot(0, this, [=]{
+        QPixmap p = QPixmap(1, 1);
+        p.fill(Qt::yellow);
+
+        QWinTaskbarButton *button = new QWinTaskbarButton(this);
+        button->setWindow(windowHandle());
+        button->setOverlayIcon(p);
+
+        QWinTaskbarProgress *progress = button->progress();
+        progress->setVisible(true);
+        progress->setValue(50);
+    });
 #endif
 }
 
@@ -127,6 +150,13 @@ void MainWindow::closeEvent(QCloseEvent *event)
         event->ignore();
 }
 
+bool MainWindow::nativeEvent(const QByteArray &eventType, void *message, long *result)
+{
+    //Listen for native events
+
+    return QGoodWindow::nativeEvent(eventType, message, result);
+}
+
 bool MainWindow::event(QEvent *event)
 {
     switch (event->type())
@@ -135,11 +165,19 @@ bool MainWindow::event(QEvent *event)
     case QEvent::Show:
     case QEvent::WindowStateChange:
     {
+        QString border_str = "none";
 #ifndef Q_OS_MAC
-        frame->setStyleSheet(m_frame_style.arg(m_color_str)
-                             .arg((isFullScreen() || isMaximized()) ? "none" : "1px solid #1883D7"));
+        if (windowState().testFlag(Qt::WindowNoState))
+        {
+            if (isActiveWindow())
+                border_str = "1px solid #1883D7";
+            else
+                border_str = "1px solid #AAAAAA";
+        }
 #endif
-        title_bar->setMaximized(isFullScreen() || isMaximized());
+        frame->setStyleSheet(m_frame_style.arg(m_color_str).arg(border_str));
+
+        title_bar->setMaximized(isMaximized());
 
         title_bar->setVisible(!isFullScreen());
 
@@ -153,8 +191,8 @@ bool MainWindow::event(QEvent *event)
     case QEvent::WindowActivate:
     {
 #ifndef Q_OS_MAC
-        frame->setStyleSheet(m_frame_style.arg(m_color_str)
-                             .arg((isFullScreen() || isMaximized()) ? "none" : "1px solid #1883D7"));
+        QString border_str = (windowState().testFlag(Qt::WindowNoState) ? "1px solid #1883D7" : "none");
+        frame->setStyleSheet(m_frame_style.arg(m_color_str).arg(border_str));
 #endif
         title_bar->setActive(true);
 
@@ -163,8 +201,8 @@ bool MainWindow::event(QEvent *event)
     case QEvent::WindowDeactivate:
     {
 #ifndef Q_OS_MAC
-        frame->setStyleSheet(m_frame_style.arg(m_color_str)
-                             .arg((isFullScreen() || isMaximized()) ? "none" : "1px solid #AAAAAA"));
+        QString border_str = (windowState().testFlag(Qt::WindowNoState) ? "1px solid #AAAAAA" : "none");
+        frame->setStyleSheet(m_frame_style.arg(m_color_str).arg(border_str));
 #endif
         title_bar->setActive(false);